(move) input condition component
